requirements.txt
python-dotenv==1.0.0
synapseclient==4.1.1
